implement accounting feature

+++ database/migrations/2021_01_08_044314_add_interface_options_to_user_settings.php
@@ -13,11 +13,21 @@ class AddInterfaceOptionsToUserSettings extends Migration
      */
     public function up()
     {
-        Schema::table('user_settings', function (Blueprint $table) {
-            $table->unsignedTinyInteger('list_pagination_size')->default(15);
-            $table->boolean('show_finished_items')->default(false);
-            $table->boolean('task_comments_sort_newest_first')->default(true);
-        });
+        if (!Schema::hasColumn('user_settings', 'list_pagination_size')) {
+            Schema::table('user_settings', function (Blueprint $table) {
+                $table->unsignedTinyInteger('list_pagination_size')->default(15);
+            });
+        }
+        if (!Schema::hasColumn('user_settings', 'show_finished_items')) {
+            Schema::table('user_settings', function (Blueprint $table) {
+                $table->boolean('show_finished_items')->default(false);
+            });
+        }
+        if (!Schema::hasColumn('user_settings', 'task_comments_sort_newest_first')) {
+            Schema::table('user_settings', function (Blueprint $table) {
+                $table->boolean('task_comments_sort_newest_first')->default(true);
+            });
+        }
     }
 
     /**
@@ -27,10 +37,20 @@ class AddInterfaceOptionsToUserSettings extends Migration
      */
     public function down()
     {
-        Schema::table('user_settings', function (Blueprint $table) {
-            $table->dropColumn('list_pagination_size');
-            $table->dropColumn('show_finished_items');
-            $table->dropColumn('task_comments_sort_newest_first');
-        });
+        if (Schema::hasColumn('user_settings', 'list_pagination_size')) {
+            Schema::table('user_settings', function (Blueprint $table) {
+                $table->dropColumn('list_pagination_size');
+            });
+        }
+        if (Schema::hasColumn('user_settings', 'show_finished_items')) {
+            Schema::table('user_settings', function (Blueprint $table) {
+                $table->dropColumn('show_finished_items');
+            });
+        }
+        if (Schema::hasColumn('user_settings', 'task_comments_sort_newest_first')) {
+            Schema::table('user_settings', function (Blueprint $table) {
+                $table->dropColumn('task_comments_sort_newest_first');
+            });
+        }
     }
 }
